Updated console.php (now works correctly!). Added link to didi's photo blog and also linked to codebase.

+++ console.php
@@ -17,14 +17,20 @@ $app->register(new ConsoleServiceProvider(), array(
 
 $app['console']
   ->register('photo:add')
-  ->setDefinition(array(
-    new InputArgument('day', InputArgument::VALUE_NONE, 'Which day are we processing? (Number only!)'),
-    new InputArgument('title', InputArgument::VALUE_NONE, 'What would you like the photo title to be?'),
-  ))
   ->setDescription('Generates and stores meta-data such as title, date and EXIF for a photo.')
-  ->setCode(function (InputInterface $input, OutputInterface $output) {
-    $day = $input->getArgument('day');
-    $title = $input->getArgument('title');
+  ->setCode(function (InputInterface $input, OutputInterface $output) use ($app) {
+    $dialog = $app['console']->getHelperSet()->get('dialog');
+    $day = $dialog->ask(
+      $output,
+      'Which day are we processing (Number only!): ',
+      ''
+    );
+
+    $title = $dialog->ask(
+      $output,
+      'What would you like the photo title to be? ',
+      ''
+    );
 
     if ($day == "") {
       // Calculate the number of day for today
@@ -36,32 +42,28 @@ $app['console']
     $date = date('Y-m-d');
 
     // Path to meta data file
-    $meta_file_path = __DIR__.'/data/metadata/'.str_pad($num, 3, 0, STR_PAD_LEFT).".json";
+    $meta_file_path = __DIR__.'/data/metadata/'.str_pad($day, 3, 0, STR_PAD_LEFT).".json";
 
     // Path to image file
-    $image_file_path = __DIR__.'/web/assets/'.str_pad($num, 3, 0, STR_PAD_LEFT).".json";
+    $image_file_path = __DIR__.'/web/assets/'.str_pad($day, 3, 0, STR_PAD_LEFT).".jpg";
 
-    $meta_info = exif_read_data($image_file_path);
+    if (file_exists($image_file_path))
+    {
+      $meta_info = exif_read_data($image_file_path);
+      $meta_data = array();
+      $meta_data["entry_day"] = $day;
+      $meta_data["entry_date"] = $date;
+      $meta_data["entry_title"] = strlen($title)?$title:$date;
+      $meta_data["image_exif"] = $meta_info;
+      $meta_data_json = json_encode($meta_data, JSON_PRETTY_PRINT);
 
-    $meta_data = array();
-    $meta_data["entry_day"] = $day;
-    $meta_data["entry_date"] = $date;
-    $meta_data["entry_title"] = strlen($title)?$title:$date;
-    $meta_data["image_exif"] = $meta_info;
-    $meta_data_json = json_encode($meta_data, JSON_PRETTY_PRINT);
-
-    file_put_contents($meta_file_path, $meta_data_json);
-    $output->writeln(sprintf('Photo for day #%s has been processed. See you tomorrow!', $day));
+      file_put_contents($meta_file_path, $meta_data_json);
+      $output->writeln(sprintf("\nPhoto for day #%s has been processed. See you tomorrow!", $day));
+    }
+    else {
+      $output->writeln(sprintf("\nERROR: Photo %s.jpg was not found.", str_pad($day, 3, 0, STR_PAD_LEFT)));
+    }
   })
 ;
-
-$commands = array(
-  new Command\XyzInfoCommand(),
-  new Command\XyzSnapshotCommand(),
-);
-
-foreach ($commands as $command) {
-  $app['console']->add($command);
-}
 
 $app['console']->run();
